remove strategy update to net 8

+++ OperationStacked/Entities/EquipmentStack.cs
@@ -7,10 +7,10 @@ public class EquipmentStack
 {
     [Key]
     public Guid Id { get; set; }
-    public Decimal StartWeight { get; set; }
-    public Decimal?[] InitialIncrements { get; set; }
+    public decimal StartWeight { get; set; }
+    public decimal?[] InitialIncrements { get; set; }
     public Decimal IncrementValue { get; set; }
-    public Decimal IncrementCount { get; set; }
+    public double IncrementCount { get; set; }
     public string EquipmentStackKey { get; set; }
     public Guid UserID { get; set; }
 }
