+++ OperationStacked/Requests/CreateEquipmentStackRequest.cs
@@ -5,7 +5,7 @@ public class CreateEquipmentStackRequest
     public Decimal StartWeight { get; set; }
     public Decimal?[] InitialIncrements { get; set; }
     public Decimal IncrementValue { get; set; }
-    public Decimal IncrementCount { get; set; }
+    public double IncrementCount { get; set; }
     public string? EquipmentStackKey { get; set; }
     public Guid UserID { get; set; }
